Fixing a few more things related to multitenancy. Removing duplicated code.
