+++ orchestrator/nodes/dockerComposeRemoteNode/index.js
@@ -1,12 +1,9 @@
 "use strict";
 
-var fs = require('fs');
+var RemoteNode = require("../remoteNode/index").Handler;
 
 // This should be external - but since it's bugged....
 var docker = require('../../docker/harbor-master');
-
-var dojot = require('@dojot/flow-node');
-var dispatcher = require('../../dispatcher');
 
 function makeId(length) {
   var text = "";
@@ -19,7 +16,7 @@ function makeId(length) {
 }
 
 
-class DataHandler extends dojot.DataHandlerBase {
+class DataHandler extends RemoteNode {
   constructor(image, id) {
     super();
     this.info = {
@@ -115,67 +112,6 @@ class DataHandler extends dojot.DataHandlerBase {
 
   update() {
     // TODO
-  }
-
-  init() {
-    // Fetch all meta information from newly created remote impl
-    return new Promise((resolve, reject) => {
-      dispatcher(this.target, {command: 'metadata'}).then((meta) => {
-        this.metadata = meta.payload;
-        dispatcher(this.target, { command: 'html' }).then((html) => {
-          this.html = '/tmp/' + this.target;
-          fs.writeFileSync(this.html, html.payload);
-          dispatcher(this.target, { command: 'locale', locale: 'en-US' }).then((reply) => {
-            this.locale = reply.payload;
-            return resolve();
-          }).catch((error) => { return reject(error); });
-        }).catch((error) => { return reject(error); });
-      }).catch((error) => { return reject(error); });
-    });
-  }
-
-  getNodeRepresentationPath() {
-    return this.html;
-  }
-
-  /**
-   * Returns node metadata information
-   * This may be used by orchestrator as a liveliness check
-   * @return {object} Metadata object
-   */
-  getMetadata() {
-    return this.metadata;
-  }
-
-  /**
-   * Returns object with locale data (for the given locale)
-   * @param  {[string]} locale Locale string, such as "en-US"
-   * @return {[object]}        Locale settings used by the module
-   */
-  getLocaleData() {
-    return this.locale;
-  }
-
-  handleMessage(config, message, callback) {
-    // invoke remote
-    let command = {
-      command: 'message',
-      message: message,
-      config: config,
-    };
-    dispatcher(this.target, command).then((reply) => {
-      if (reply.error) {
-        return callback(reply.error);
-      }
-
-      if (Array.isArray(reply)){
-        return callback(undefined, reply);
-      }
-
-      return callback(new Error("Invalid response received from node"));
-    }).catch((error) => {
-      callback(error);
-    });
   }
 }
